src/pyproject.toml
[tool.poetry]
name = "chainlit"
version = "0.1.1"
keywords = ['LLM', 'Agents', 'gen ai', 'chat ui', 'chatbot ui', 'langchain']
description = "A faster way to build chatbot UIs."
authors = ["Chainlit"]
license = "Apache-2.0 license"
repository = "https://github.com/Chainlit/chainlit"
readme = "README.md"
exclude = [
    "**/frontend/**/*",
]
include = [
    "**/frontend/dist/**/*",
]

[tool.poetry.scripts]
# command_name = module_for_handler : function_for_handler
chainlit = 'chainlit.cli:cli'

[tool.poetry.dependencies]
python = "^3.8"
dataclasses_json = "^0.5.7"
flask = "^2.2.3"
flask_socketio = "^5.3.3"
flask_cors = "^3.0.10"
flask-limiter = "^3.3.0"
click = "^8.1.3"
openai = "^0.27.2"
tomli = "^2.0.1"
watchdog = "^3.0.0"
watchdog_gevent = "^0.1.1"
gevent = "^22.10.2"
gevent_websocket = "^0.10.1"
pydantic = "^1.10.6"
python-graphql-client = "^0.4.3"
python-dotenv = "^1.0.0"
auth0-python = "^4.1.1"
uptrace = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
